Fix idle disk issue
The program was crashing sometimes, usually when the Time Machine disk
was idle. However, it was also observed that after launch the command
the disk woke up. Therefore this workaround tries to fix the problem
simply relaunching the command several times until a call is successful,
obviously limiting ourselves to just a few tries.